+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: epolitze <epolitze@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/17 14:10:13 by epolitze          #+#    #+#             */
-/*   Updated: 2023/11/23 13:07:38 by epolitze         ###   ########.fr       */
+/*   Updated: 2023/11/23 17:45:14 by epolitze         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,8 +22,8 @@ int		ft_putchar(char c);
 int		ft_putstr(char *str);
 int		ft_putaddress(void *ptr, int wcount);
 int		ft_putnbr(long long n, int wcount);
-int		ft_putunbr(unsigned long long n, int wcount);
+int		ft_putunbr(unsigned long int n, int wcount);
 size_t	ft_strlen(char *str);
-int		ft_puthex(long long nb, int lower, int wcount);
+int		ft_puthex(unsigned long int nb, int lower, int wcount);
 
 #endif
